improvement: simplify elements
