+++ packages/jaspr/lib/src/framework/render_object.dart
@@ -22,27 +22,9 @@ abstract class RenderObject {
   void remove();
 }
 
-abstract class MultiChildRenderObjectElement extends MultiChildElement with RenderObjectElement {
-  MultiChildRenderObjectElement(super.component);
-
-  @override
-  void update(covariant Component newComponent) {
-    super.update(newComponent);
-    _dirty = true;
-    rebuild();
-  }
-}
-
-abstract class SingleChildRenderObjectElement extends SingleChildElement with RenderObjectElement {
-  SingleChildRenderObjectElement(super.component);
-
-  @override
-  void update(covariant Component newComponent) {
-    super.update(newComponent);
-    _dirty = true;
-    rebuild();
-  }
-}
+abstract class BuildableRenderObjectElement = BuildableElement with RenderObjectElement;
+abstract class ProxyRenderObjectElement = ProxyElement with RenderObjectElement;
+abstract class LeafRenderObjectElement = LeafElement with RenderObjectElement;
 
 mixin RenderObjectElement on Element {
   void updateRenderObject();
@@ -51,12 +33,12 @@ mixin RenderObjectElement on Element {
   RenderObject? _renderObject;
 
   @override
-  void _firstBuild() {
+  void didMount() {
     if (_renderObject == null) {
       _renderObject = _parentRenderObjectElement!.renderObject.createChildRenderObject();
       updateRenderObject();
     }
-    super._firstBuild();
+    super.didMount();
   }
 
   @override
